Changed many things to evaluate sizes in types (eg int bitsize) as early as possible, because otherwise it was taking too long when a lot of little additions and Max operations gathered up.
Unroll.hs: enable counting up as well as down in for loops.

CircGen.hs: generated vars (temp(i)) are now not selected on at the end of a
conditional branch - they should not need to persist non-locally (I hope!)
Also, removed the call to 'tr' in showCct - profiling showed it takes a lot
of time.