Finally running heap.sfdl correctly!!
Expanded "print" support, including in the runtime. Arranged to have printed
values be placed at the print gate, so they are subsequently used from
there, and the print gates is thus forced to appear in a reasonable
location.

Unroll.hs: stuuupid bug in unrolling for-loops counting
backwards - was counting up instead of down as I forgot to reverse
the list.